feat: append outfile and chores: change filestructure and fts with norm in mind
